# backend/openapi.yaml
openapi: 3.0.3
info:
  title: SurveyLauncher API
  description: |
    Comprehensive REST API for SurveyLauncher mobile device management platform.

    The SurveyLauncher backend provides a comprehensive REST API with two distinct interfaces:
    1. **Mobile App API** (`/api/v1/*`) - For Android device authentication and field operations
    2. **Web Admin API** (`/api/web-admin/*`) - For administrative and management functions

    The API follows a consistent JSON format with proper error handling and response structures,
    implementing enterprise-grade role-based access control (RBAC) with nine specialized roles.

    ## Authentication
    Most endpoints require JWT token authentication via the `Authorization` header:
    ```
    Authorization: Bearer <access_token>
    ```

    ## Role-Based Access Control
    The API implements enterprise-grade RBAC with nine specialized roles:

    ### 🟢 App + Web Roles (Hybrid Access)
    - `TEAM_MEMBER` - Field users
    - `FIELD_SUPERVISOR` - Frontline supervisors
    - `REGIONAL_MANAGER` - Regional leadership

    ### 🔵 Web Admin Only Roles (Web-Only Access)
    - `SYSTEM_ADMIN` - Full system administrator
    - `SUPPORT_AGENT` - Customer support
    - `AUDITOR` - Compliance auditor
    - `DEVICE_MANAGER` - Device management
    - `POLICY_ADMIN` - Policy administrator
    - `NATIONAL_SUPPORT_ADMIN` - National-level support
  version: 1.0.0
  contact:
    name: SurveyLauncher Team
    email: support@surveylauncher.com
  license:
...
servers:
...
paths:
...
  /api/v1/teams:
    get:
      tags: [Teams]
      summary: List teams
      description: List teams with pagination and search
      operationId: listTeams
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/IsActive'
      responses:
        '200':
          description: Teams retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamsListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Teams]
      summary: Create team
      description: Create a new team
      operationId: createTeam
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeamRequest'
      responses:
        '201':
          description: Team created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/teams/{id}:
    get:
      tags: [Teams]
      summary: Get team by ID
      description: Get a specific team by ID
      operationId: getTeam
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TeamId'
      responses:
        '200':
          description: Team retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags: [Teams]
      summary: Update team
      description: Update team information
      operationId: updateTeam
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TeamId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTeamRequest'
      responses:
        '200':
          description: Team updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Teams]
      summary: Delete team
      description: Delete a team (soft delete)
      operationId: deleteTeam
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TeamId'
      responses:
        '200':
          description: Team deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  # Devices endpoints
...
  /api/v1/users:
    get:
      tags: [Users]
      summary: List users
      description: List users with pagination and search
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/TeamId'
        - $ref: '#/components/parameters/Role'
        - $ref: '#/components/parameters/IsActive'
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Users]
      summary: Create user
      description: Create a new user with PIN
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/users/{id}:
    get:
      tags: [Users]
      summary: Get user by ID
      description: Get a specific user by ID
      operationId: getUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags: [Users]
      summary: Update user
      description: Update user information
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Users]
      summary: Delete user
      description: Delete a user (soft delete)
      operationId: deleteUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # Policy endpoint
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token for mobile app authentication
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: HTTP-only cookie for web admin authentication
  parameters:
    Page:
      in: query
      name: page
      description: Page number for pagination
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      in: query
      name: limit
      description: Items per page (max 100)
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    Search:
      in: query
      name: search
      description: Search term
      required: false
      schema:
        type: string
    IsActive:
      in: query
      name: is_active
      description: Filter by active status
      required: false
      schema:
        type: boolean
    TeamId:
      in: path
      name: id
      description: Team UUID
      required: true
      schema:
        type: string
        format: uuid
...
    UserId:
      in: path
      name: id
      description: User UUID
      required: true
      schema:
        type: string
        format: uuid
    Role:
      in: query
      name: role
      description: Filter by role
      required: false
      schema:
        type: string
        enum: [TEAM_MEMBER, FIELD_SUPERVISOR, REGIONAL_MANAGER, SYSTEM_ADMIN, SUPPORT_AGENT, AUDITOR, DEVICE_MANAGER, POLICY_ADMIN, NATIONAL_SUPPORT_ADMIN]
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        message:
          type: string
          example: Operation successful
    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        error:
          $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      properties:
        code:
          type: string
          example: ERROR_CODE
        message:
          type: string
          example: Human-readable error description
        request_id:
          type: string
          example: unique-request-id
...
    TeamResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        team:
          $ref: '#/components/schemas/Team'
    Team:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        timezone:
          type: string
        state_id:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TeamsListResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        teams:
          type: array
          items:
            $ref: '#/components/schemas/Team'
        pagination:
          $ref: '#/components/schemas/Pagination'
    CreateTeamRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
        timezone:
          type: string
          default: UTC
        stateId:
          type: string
    UpdateTeamRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        timezone:
          type: string
...
    UserResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        team_id:
          type: string
          format: uuid
        code:
          type: string
        display_name:
          type: string
        email:
          type: string
          nullable: true
        role:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UsersListResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
        pagination:
          $ref: '#/components/schemas/Pagination'
    CreateUserRequest:
      type: object
      required:
        - teamId
        - code
        - displayName
        - pin
      properties:
        teamId:
          type: string
          format: uuid
        code:
          type: string
        displayName:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        role:
          type: string
        pin:
          type: string
          minLength: 4
    UpdateUserRequest:
      type: object
      properties:
        displayName:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        role:
          type: string
        isActive:
          type: boolean
        pin:
          type: string
          minLength: 4
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        pages:
          type: integer
          minimum: 0
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            ok: false
            error:
              code: MISSING_FIELDS
              message: Required fields are missing
              request_id: unique-request-id
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            ok: false
            error:
              code: UNAUTHORIZED
              message: Authentication required
              request_id: unique-request-id
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            ok: false
            error:
              code: NOT_FOUND
              message: Resource not found
              request_id: unique-request-id
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            ok: false
            error:
              code: CONFLICT
              message: Resource already exists
              request_id: unique-request-id
...